+++ bagua/torch_api/contrib/__init__.py
@@ -1,4 +1,4 @@
-from .fused_optimizer import FusedOptimizer  # noqa: F401
+from .fuse.optimizer import FusedOptimizer  # noqa: F401
 from .load_balancing_data_loader import (  # noqa: F401
     LoadBalancingDistributedSampler,
     LoadBalancingDistributedBatchSampler,
